sample_api_server/requirements.txt
Django==1.10.2
djangorestframework==3.4.6